requirements.txt
Django
djangorestframework
numpy
librosa
tensorflow
  
